+++ R/countreg-support.R
@@ -71,7 +71,8 @@ emm_basis.zeroinfl = function(object, trms, xlev, grid,
             lp = as.numeric(X %*% bhat + .get.offset(trms, grid))
             lnk = make.link(misc$tran)
             bhat = lnk$linkinv(lp)
-            delta = .diag(lnk$mu.eta(lp)) %*% X
+#            delta = .diag(lnk$mu.eta(lp)) %*% X
+            delta = sweep(X, 1, lnk$mu.eta(lp), "*")
             V = delta %*% tcrossprod(V, delta)
             X = diag(1, length(bhat))
             misc = list(offset.mult = 0)
@@ -96,14 +97,16 @@ emm_basis.zeroinfl = function(object, trms, xlev, grid,
         mu2prime = stats::make.link(object$link)$mu.eta(lp2)
         
         if(mode == "response") {
-            delta = .diag(mu1) %*% cbind(.diag(1 - mu2) %*% X1, .diag(-mu2prime) %*% X2)
+            delta = cbind(sweep(X1, 1, mu1 * (1 - mu2), "*"), 
+                          sweep(X2, 1, -mu1 * mu2prime, "*"))
             bhat = (1 - mu2) * mu1
         }
         else { # mode = "prob0"
-            p0 = 1 - .prob.gt.0(object$dist, mu1, object$theta)
-            dp0 = - .dprob.gt.0(object$dist, mu1, object$theta, "log", lp1)
+            tmp = .zi.support(mu1, object$theta, .make.p0(object$dist))
+            p0 = tmp[1, ]; dp0 = tmp[2, ]
             bhat = (1 - mu2) * p0 + mu2
-            delta = cbind(.diag((1 - mu2) * dp0) %*% X1, .diag(mu2prime * (1 - p0)) %*% X2)
+            delta = cbind(sweep(X1, 1, (1 - mu2) * dp0, "*"),
+                          sweep(X2, 1, mu2prime * (1 - p0), "*"))
         }
         V = delta %*% tcrossprod(.pscl.vcov(object, model = "full", ...), delta)
         X = diag(1, length(bhat))
@@ -154,14 +157,14 @@ emm_basis.hurdle = function(object, trms, xlev, grid,
             lp = as.numeric(X %*% bhat + .get.offset(trms, grid))
             lnk = make.link(misc$tran)
             bhat = lnk$linkinv(lp)
-            if (mode != "prob0") {
-               delta = .diag(lnk$mu.eta(lp)) %*% X
+            mult = lnk$mu.eta(lp)
+            if(mode == "prob0") {
+                shape = 
+                tmp = .zi.support(bhat, object$theta["zero"], .make.p0(object$dist$zero))
+                bhat = tmp[1, ]
+                mult = mult * tmp[2, ]
             }
-            else {
-                bhat = 1 - .prob.gt.0(object$dist$zero, bhat, object$theta["zero"])
-                db = - .dprob.gt.0(object$dist$zero, bhat, object$theta["zero"], misc$tran, lp)
-                delta = .diag(db) %*% X
-            }
+            delta = sweep(X, 1, mult, "*")
             V = delta %*% tcrossprod(V, delta)
             X = diag(1, length(bhat))
             misc = list(offset.mult = 0)
@@ -173,10 +176,8 @@ emm_basis.hurdle = function(object, trms, xlev, grid,
         contr1 = object$contrasts[["count"]]
         X1 = model.matrix(trms1, m, contrasts.arg = contr1)
         b1 = coef(object, model = "count")
-        mu1 = as.numeric(exp(X1 %*% b1 + off1))
+        mu1 = mu1prime = as.numeric(exp(X1 %*% b1 + off1))
         theta1 = object$theta["count"]
-        p1 = .prob.gt.0(object$dist$count, mu1, theta1)
-        dp1 = .dprob.gt.0(object$dist$count, mu1, theta1, "", 0) # binomial won't happen
 
         trms2 = delete.response(terms(object, model = "zero"))
         off2 = .get.offset(trms2, grid)
@@ -188,18 +189,26 @@ emm_basis.hurdle = function(object, trms, xlev, grid,
                      binomial = object$linkinv(lp2),
                      exp(lp2)  )
         theta2 = object$theta["zero"]
-        p2 = .prob.gt.0(object$dist$zero, mu2, theta2)
-        dp2 = .dprob.gt.0(object$dist$zero, mu2, theta2, object$link, lp2)
-
+        mu2prime = switch(object$dist$zero, 
+                          binomial = stats::make.link("logit")$mu.eta(lp2),
+                          exp(lp2) )
         if (mode == "response") {
-            bhat = p2 * mu1 / p1
-            delta = cbind(.diag(bhat*(1 - mu1 * dp1 / p1)) %*% X1,
-                          .diag(mu1 * dp2 / p1) %*% X2)
+            tmp = .hurdle.support(mu1, theta1, .make.p0(object$dist$count), \(mu, shape) mu, 
+                                  mu2, theta2, .make.p0(object$dist$zero))
+            bhat = tmp[1, ]
+            delta = cbind(sweep(X1, 1, tmp[2, ] * mu1prime, "*"),
+                          sweep(X2, 1, tmp[3, ] * mu2prime, "*"))
         }
         else {  ## mode == "zero"
+            tmp1 = .zi.support(mu1, theta1, .make.p0(object$dist$count))
+            tmp2 = .zi.support(mu2, theta2, .make.p0(object$dist$zero))
+            p1 = 1 - tmp1[1, ]; dp1 = -tmp1[2, ]
+            p2 = 1 - tmp2[1, ]; dp2 = -tmp2[2, ]
             bhat = p2 / p1
-            delta = cbind(.diag(-p2 * dp1 / p1^2) %*% X1,
-                          .diag(dp2 / p1) %*% X2)
+            # delta = cbind(.diag(-p2 * dp1 / p1^2) %*% X1,
+            #               .diag(dp2 / p1) %*% X2)
+            delta = cbind(sweep(X1, 1, -p2 * dp1, "*"),
+                          sweep(X2, 1, dp2 / p1, "*"))
         }
         V = delta %*% tcrossprod(.pscl.vcov(object, model = "full", ...), delta)
         X = .diag(1, length(bhat))
@@ -213,25 +222,86 @@ emm_basis.hurdle = function(object, trms, xlev, grid,
          dffun = dffun, dfargs = dfargs, misc = misc)
 }
 
-# utility for prob (Y > 0 | dist, mu, theta)
-.prob.gt.0 = function(dist, mu, theta) {
+
+#' @rdname extending-emmeans
+#' @order 93
+#'
+#' @param cmu,zmu Vector of back-transformed estimates for the count and zero model, resp 
+#' @param cshape,zshape Shape parameter for the count and zero model, resp 
+#' @param cp0,zp0 Function of \code{(mu, shape)} for computing Prob(Y = 0)
+#'        for the count and zero model, resp
+#' @param cmean Function of \code{(mu, shape)} for computing the mean of the
+#'        count model. Typically, this just returns \code{mu}
+#'
+#' @return \code{.hurdle.support} returns a matrix with 3 rows containing the
+#'       estimated mean responses and the differentials wrt \code{cmu} and \code{zmu},
+#'       resp. 
+#' @export
+.hurdle.support = function(cmu, cshape, cp0, cmean, 
+                           zmu, zshape, zp0) {
+    if (is.null(cshape) || is.na(cshape)) cshape = 1
+    if (is.null(zshape) || is.na(zshape)) zshape = 1
+    mfcn = function(x) (1 - zp0(x[2], zshape)) * cmean(x[1], cshape) / 
+        (1 - cp0(x[1], cshape))
+    result = sapply(seq_along(cmu), function(i) {
+        x = c(cmu[i], zmu[i])
+        d = numDeriv::jacobian(mfcn, x)
+        c(mfcn(x), d)
+    })
+    rownames(result) = c("mean", "dcmu", "dzmu")
+    result
+}
+
+#' @rdname extending-emmeans
+#' @order 94
+#' @param mu,shape,p0 See parameters \code{zmu,zshape,zp0} above
+#' @return \code{.zi.support} returns a matrix with 2 rows containing the
+#'       estimated probabilities of 0 and the differentials wrt \code{mu}.
+#'       See the section on hurdle and zero-inflated models.
+#' @section Support for Hurdle and Zero-inflated models:
+#'   The functions \code{.hurdle.support} and \code{.zi.support} help facilitate
+#'   calculations needed to estimate the mean response (count model and zero model
+#'   combined) of these models. \code{.hurdle.support} returns a matrix of three rows.
+#'   The first is the estimated mean for a hurdle model, and the 2nd and 3rd rows are
+#'   differentials for the count and zero models, which needed for delta-method
+#'   calculations. To use these, regard the \code{@linfct} slot as comprising
+#'   two sets of columns, for the count and zero models respectively. To do
+#'   the delta method calculations, multiply the rows of the count part by its 
+#'   differentials times \code{link$mu.eta} evcaluated at that part of the linear predictor.
+#'   Do the same for the zero part, using its differentials and \code{mu.eta}.
+#'   If the resulting matrix is \bold{A}, then the covariance of the mean response
+#'   is \bold{AVA'} where \bold{V}is the \code{@V} slot of the object.
+#'   
+#'   The function \code{zi.support} works the same way, only it is much simpler,
+#'   and is used to estimate the probability of 0 and its differential for either 
+#'   part of a zero-inflated model or hurdle model.
+#'   
+#'   See the code for \code{emm_basis.zeroinfl} and \code{emm_basis.hurdle}
+#'   for how these are used with models fitted by the \pkg{pscl} package.
+#' @export
+#'
+.zi.support = function(mu, shape, p0) {
+    if (is.null(shape) || is.na(shape)) shape = 1
+    pfcn = function(x) p0(x, shape)
+    result = sapply(seq_along(mu), function(i) {
+        c(pfcn(mu[i]), numDeriv::jacobian(pfcn, mu[i]))
+    })
+    rownames(result) = c("p0", "dmu")
+    result
+}
+
+
+.make.p0 = function(dist) {
     switch(dist,
-       binomial = mu,
-       poisson = 1 - exp(-mu),
-       negbin = 1 - (theta / (mu + theta))^theta,
-       geometric = 1 - 1 / (1 + mu)
+           bernoulli = \(mu, shape) 1 - mu,
+           binomial = \(mu, shape) (1 - mu)^shape,
+           poisson = \(mu, shape) exp(-mu),
+           negbin = \(mu, shape) (shape / (mu + shape))^shape,
+           geometric = \(mu, shape) 1 / (1 + mu)
     )
 }
 
-# utility for d/d(eta) prob (Y > 0 | dist, mu, theta)
-.dprob.gt.0 = function(dist, mu, theta, link, lp) {
-    switch(dist,
-       binomial = stats::make.link(link)$mu.eta(lp),
-       poisson = mu * exp(-mu),
-       negbin = mu * (theta /(mu + theta))^(1 + theta),
-       geometric = mu / (1 + mu)^2  
-    )
-}
+
 
 # special version of .my.vcov that accepts (and requires!) model argument
 .pscl.vcov = function(object, model, vcov. = stats::vcov, ...) {
@@ -242,90 +312,3 @@ emm_basis.hurdle = function(object, trms, xlev, grid,
     vcov.
 }
 
-# Explanatory notes for hurdle models
-# -----------------------------------
-#     We have a linear predictor eta = X%*%beta + offset
-#     mu = h(eta) where h is inverse link (usually exp but not always)
-#     Define p = P(Y > 0 | mu). This comes out to...
-#         binomial: mu
-#         poisson: 1 - exp(-mu)
-#         negbin: 1 - (theta/(mu+theta))^theta
-#         geometric: 1 - 1/(mu+1)
-#     Define dp = dp/d(eta). Note - when h(mu)=exp(mu) we have dp = mu*dp/d(mu)
-#         binomial: h'(eta)
-#         poisson: mu*exp(-mu)
-#         negbin: mu*(theta/(mu+theta))^(theta+1)
-#         geometric: mu/(mu+1)^2
-#     
-#     This gives us what we need to find the estimates and apply the delta method
-#     In the code we index these notations with 1 (count model) and 2 (zero model)
-#     And we treat theta1 and theta2 as constants
-#
-#!!! In theory, above seems correct, and estimates match those from predict.hurdle.
-#!!! But SEs don't seem right. 
-#!!! They do seem right though if I omit the factor of mu in dp
-#!!! when link is log
-
-
-### Simulation-based approach for ZI and hurdle models
-### This returns a list of ther same form as an emm_basis() method
-##   X is model matrix for both models (X.count | X.zi)
-##   ncoef.c # cols in X.count
-##   bhat is all regression coefs -- OR a matrix w/ posterior samples
-##   V is combined vcov matrix (ignored if bhat is a matrix)
-##   links is vector (or list) of links for the 2 parts of the model
-##   fams is list of family names
-##   hurdle is a named list(famc, thetac, famz, thetaz) or empty if ZI
-##   N.sim is # simulations desired
-##   keep.sim set to TRUE to return sims in post.beta
-##   df is d.f. to return
-## NOTE if bhat is a matrix, keep.sim is set to TRUE and N.sim is ignored
-#' @export
-.zi.simulate = function(X, ncoef.c, bhat, V, links, hurdle = list(), 
-                        N.sim = 1000, keep.sim = FALSE,
-                        df = Inf, misc = list()) 
-{
-    if(is.matrix(bhat)) {
-        keep.sim = TRUE
-        B = bhat
-        bhat = NA
-    }
-    else {
-        if (length(bhat) != ncol(V))
-            stop("Non-estimable cases not yet supported for zero-inflation calculations")
-        B = mvtnorm::rmvnorm(N.sim, bhat, V)
-    }
-    back.tran = function(idx, link) {
-        W = B[, idx] %*% t(X[, idx])
-        if (is.character(link))
-            link = make.link(link)
-        link$linkinv(W)
-    }
-    if(!is.na(bhat[1]))
-        B = rbind(bhat, B) # put our pt est in 1st row
-    ic = seq_len(ncoef.c)
-    iz = setdiff(seq_len(ncol(B)), ic)
-    C = back.tran(ic, links[[1]])
-    Z = back.tran(iz, links[[2]])
-    if (length(hurdle) >= 4)
-        R = C * .prob.gt.0(hurdle$famz, Z, hurdle$thetaz) / 
-                .prob.gt.0(hurdle$famc, C, hurdle$thetac)
-    else
-        R = (1 - Z) * C
-    if(!is.na(bhat[1])){
-        bhat = R[1, ]
-        R = R[-1, ]
-    }
-    else
-        bhat = apply(R, 2, mean)
-    V = cov(R)
-    dffun = function(k, dfargs) dfargs$df
-    dfargs = list(df = df)
-    if(is.null(misc$est.name)) 
-        misc$estName = "emmean"
-    if(!keep.sim)
-        post.beta = matrix(NA)
-    list(X = diag(length(bhat)), bhat = bhat, nbasis = estimability::all.estble, V = V, 
-         dffun = dffun, dfargs = dfargs, misc = misc, post.beta = R)
-}
-